// types.ts
export type TransactionType = 'credit' | 'debit' | 'pix' | 'boleto';

export interface Transaction {
  id: string;
  description: string;
  amount: number;
  date: string;
  bankName: string;
  category: string;
  type: TransactionType;
  icon: React.FC<{className?: string}>;
  notes?: string;
}

export type InvoiceStatus = 'COMPLETED' | 'PROCESSING' | 'FAILED';

export interface Invoice {
  id: string;
  fileName: string;
  uploadDate: string;
  status: InvoiceStatus;
}

export type UserRole = 'FREE' | 'PREMIUM';

export interface User {
  name: string;
  role: UserRole;
}
// ...
export type InsightType = 'SAVINGS_OPPORTUNITY' | 'UNUSUAL_SPENDING' | 'CASH_FLOW_FORECAST';

export interface FinancialInsight {
  id: string;
  type: InsightType;
  title: string;
  description: string;
  actionText: string;
}